+++ otelcollector/configmapparser/tomlparser-scrape-interval.rb
@@ -26,7 +26,7 @@ MATCHER = /^((([0-9]+)y)?(([0-9]+)w)?(([0-9]+)d)?(([0-9]+)h)?(([0-9]+)m)?(([0-9]
 @windowskubeproxyScrapeInterval = "30s"
 @prometheusCollectorHealthInterval = "30s"
 @podannotationScrapeInterval = "30s"
-@kappiebasicScrapeInterval = "30s"
+@retinabasicScrapeInterval = "30s"
 
 # Use parser to parse the configmap toml file to a ruby structure
 def parseConfigMap
@@ -192,20 +192,20 @@ def populateSettingValuesFromConfigMap(parsedConfig)
       ConfigParseErrorLogger.log(LOGGING_PREFIX, "windowskubeproxyScrapeInterval override not specified in configmap")
     end
 
-    kappiebasicScrapeInterval = parsedConfig[:kappiebasic]
-    if !kappiebasicScrapeInterval.nil?
-      matched = MATCHER.match(kappiebasicScrapeInterval)
+    retinabasicScrapeInterval = parsedConfig[:retinabasic]
+    if !retinabasicScrapeInterval.nil?
+      matched = MATCHER.match(retinabasicScrapeInterval)
       if !matched
         # set default scrape interval to 30s if its not in the proper format
-        kappiebasicScrapeInterval = "30s"
-        @kappiebasicScrapeInterval = kappiebasicScrapeInterval
-        ConfigParseErrorLogger.log(LOGGING_PREFIX, "Incorrect regex pattern for duration, set default scrape interval to 30s for kappie")
+        retinabasicScrapeInterval = "30s"
+        @retinabasicScrapeInterval = retinabasicScrapeInterval
+        ConfigParseErrorLogger.log(LOGGING_PREFIX, "Incorrect regex pattern for duration, set default scrape interval to 30s for retina")
       else
-        @kappiebasicScrapeInterval = kappiebasicScrapeInterval
-        ConfigParseErrorLogger.log(LOGGING_PREFIX, "Using configmap scrape settings for kappiebasicScrapeInterval")
+        @retinabasicScrapeInterval = retinabasicScrapeInterval
+        ConfigParseErrorLogger.log(LOGGING_PREFIX, "Using configmap scrape settings for retinabasicScrapeInterval")
       end
     else
-      ConfigParseErrorLogger.log(LOGGING_PREFIX, "kappiebasicScrapeInterval override not specified in configmap")
+      ConfigParseErrorLogger.log(LOGGING_PREFIX, "retinabasicScrapeInterval override not specified in configmap")
     end
 
     prometheusCollectorHealthInterval = parsedConfig[:prometheuscollectorhealth]
@@ -269,7 +269,7 @@ intervalHash["WINDOWSEXPORTER_SCRAPE_INTERVAL"] = @windowsexporterScrapeInterval
 intervalHash["WINDOWSKUBEPROXY_SCRAPE_INTERVAL"] = @windowskubeproxyScrapeInterval
 intervalHash["PROMETHEUS_COLLECTOR_HEALTH_SCRAPE_INTERVAL"] = @prometheusCollectorHealthInterval
 intervalHash["POD_ANNOTATION_SCRAPE_INTERVAL"] = @podannotationScrapeInterval
-intervalHash["KAPPIEBASIC_SCRAPE_INTERVAL"] = @kappiebasicScrapeInterval
+intervalHash["retinaBASIC_SCRAPE_INTERVAL"] = @retinabasicScrapeInterval
 
 if !file.nil?
   # Close file after writing scrape interval list hash